Site/src/test/javascript/spec/app/entities/estado-candidatura/estado-candidatura.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { EstadoCandidaturaService } from 'app/entities/estado-candidatura/estado-candidatura.service';
import { IEstadoCandidatura, EstadoCandidatura } from 'app/shared/model/estado-candidatura.model';

describe('Service Tests', () => {
  describe('EstadoCandidatura Service', () => {
    let injector: TestBed;
    let service: EstadoCandidaturaService;
    let httpMock: HttpTestingController;
    let elemDefault: IEstadoCandidatura;
    let expectedResult: IEstadoCandidatura | IEstadoCandidatura[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(EstadoCandidaturaService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new EstadoCandidatura(0, 'AAAAAAA', false);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a EstadoCandidatura', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new EstadoCandidatura()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a EstadoCandidatura', () => {
        const returnedFromService = Object.assign(
          {
            nombre: 'BBBBBB',
            porDefecto: true,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of EstadoCandidatura', () => {
        const returnedFromService = Object.assign(
          {
            nombre: 'BBBBBB',
            porDefecto: true,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a EstadoCandidatura', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
